# Optimised websocket message sending