Currently theres a lot of repeated work in the preparing and sending of messages. This should be easy to improve.

Some numbers:
```
Master:

create x 1:    8.08 ms, 8.08 ms/item, 123.67 items/sec
create x 16:   10.96 ms, 0.68 ms/item, 1459.72 items/sec
create x 256:  88.39 ms, 0.34 ms/item, 2896.18 items/sec
create x 4096: 1245.62 ms, 0.30 ms/item, 3288.29 items/sec
delete x 1:    7.63 ms, 7.63 ms/item, 130.90 items/sec
delete x 16:   11.63 ms, 0.72 ms/item, 1374.80 items/sec
delete x 256:  89.42 ms, 0.34 ms/item, 2862.83 items/sec
delete x 4096: 1259.40 ms, 0.30 ms/item, 3252.33 items/sec

Optimised:

create x 1:    8.43 ms, 8.43 ms/item, 118.53 items/sec
create x 16:   9.29 ms, 0.58 ms/item, 1721.72 items/sec
create x 256:  8.06 ms, 0.03 ms/item, 31730.29 items/sec
create x 4096: 9.55 ms, 0.00 ms/item, 428900.52 items/sec
delete x 1:    3.92 ms, 3.92 ms/item, 254.58 items/sec
delete x 16:   4.30 ms, 0.26 ms/item, 3716.60 items/sec
delete x 256:  4.17 ms, 0.01 ms/item, 61273.33 items/sec
delete x 4096: 10.16 ms, 0.00 ms/item, 402911.66 items/sec
```

**Note:** This does not include the actual time to send the reply, only do all of the work to decide to. That sending process is likely pretty significant, but also outside of our control.

The message here is that:

 * In normal circumstances we aren't going to see much performance difference
 * ... but our scaling is _much_ better if we need to send a lot of notifications (close to constant time for any reasonable number)